changed internal rep of characteristic from int to long.
Some computations cannot handle such large characteristic, e.g.
factorization of polynomials.  These still need better error
messages when given inappropriate input

+++ M2/Macaulay2/e/tower.hpp
@@ -21,7 +21,7 @@ class Tower : public Ring
 protected:
   Tower() {}
 
-  bool initialize(int charac0, M2_ArrayString names0, const VECTOR(ring_elem) &extensions);
+  bool initialize(long charac0, M2_ArrayString names0, const VECTOR(ring_elem) &extensions);
 public:
   virtual ~Tower();
